Add the GeoJSON endpoint to the OpenAPI documentation
