+++ server/src/routes/survey.routes.ts
@@ -21,7 +21,12 @@ import {
   userCanEditSurvey,
   userCanViewSurvey,
 } from '@src/application/survey';
-import { ensureAuthenticated, ensureSurveyGroupAccess } from '@src/auth';
+import {
+  ensureAuthenticated,
+  ensurePublicationAccess,
+  ensureSurveyGroupAccess
+} from '@src/auth';
+import { getGeometryDBEntriesAsGeoJSON } from '@src/application/answer';
 import { ForbiddenError } from '@src/error';
 import { Router } from 'express';
 import asyncHandler from 'express-async-handler';
@@ -470,6 +475,38 @@ router.get(
 
     const submissions = await getSubmissionsForSurvey(surveyId);
     res.json(submissions);
+  }),
+);
+
+/**
+ * Get submissions for the map questions as vector layers
+ */
+router.get(
+  '/:id/submissions/map',
+  validateRequest([
+    param('id').isNumeric().toInt().withMessage('ID must be a number'),
+    query('question').toArray()
+  ]),
+  ensurePublicationAccess(),
+  asyncHandler(async (req, res) => {
+    const surveyId = Number(req.params.id);
+
+    const layers = await getGeometryDBEntriesAsGeoJSON(surveyId) ?? {};
+    const layerArr = Object.entries(layers)
+
+    // Filter layers by question number(s), possibly received as query parameters
+    const questionsToReturn = (req.query.question as string[])
+    .map(q => parseInt(q))
+    .filter(q => !isNaN(q) && q > 0 && q <= layerArr.length);
+
+    // To do: Use Object.fromEntries instead of reduce. Currently it gives a Typescript error
+    res.json(
+      !questionsToReturn.length ? layers : questionsToReturn.reduce((filtered, i) => {
+        const layer = layerArr[i-1];
+        filtered[layer[0]] = layer[1];
+        return filtered
+      }, {})
+    );
   }),
 );
